backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
numpy==2.1.3
llama-index-core==0.12.0
llama-index-llms-openrouter==0.3.0
python-multipart==0.0.12
